[Refactor] #1 - 패키지 구조 도메인형 -> 계층형 변경
