+++ Assignment/sopt34-week1-assignment/src/main/java/org/example/Main.java
@@ -1,10 +1,9 @@
 package org.example;
 
-import org.example.account.Account;
-import org.example.account.AccountService;
-import org.example.account.CheckingAccount;
-import org.example.account.CheckingAccountService;
-import org.example.bank.BankController;
+import org.example.service.AccountService;
+import org.example.domain.CheckingAccount;
+import org.example.service.CheckingAccountService;
+import org.example.controller.BankController;
 
 public class Main {
     public static void main(String[] args) {
