+++ Assignment/sopt34-week1-assignment/src/main/java/org/example/controller/BankController.java
@@ -1,6 +1,6 @@
-package org.example.bank;
+package org.example.controller;
 
-import org.example.account.AccountService;
+import org.example.service.AccountService;
 
 import javax.naming.InsufficientResourcesException;
 import java.util.Scanner;
