+++ Assignment/sopt34-week1-assignment/src/main/java/org/example/domain/Account.java
@@ -1,4 +1,4 @@
-package org.example.account;
+package org.example.domain;
 
 import javax.naming.InsufficientResourcesException;
 import java.util.Random;
